refactor: remove VR glsl and rename shader modes
